Use www.disjoint.ca as the base url
The primary reason for this is that one is not allowed to use a CNAME
record in an apex domain (i.e. disjoint.ca). CNAME records are only
allowed on a non-apex domain.

And since cloudfront does not play so well with non-cname pointers,
attempting to point disjoint.ca as a CNAME record to the cloudfront
leaf wont work.

TODO: See if it's worth using Route 53 alias records for this purpose.