Use the Twitter Web Intents functionality for the "tweet" button
This is a all around better solution since it does the right thing
depending on context.

For example, on an Android, the corresponding javascript will take care
of launching the Twitter mobile app (assuming a person has it installed)
and pre-fill it as if it were a desktop client.

More Info at https://dev.twitter.com/web/intents